inserter/app/requirements.txt
Flask
psycopg2-binary
flask_cors
requests